feat: implement Our Sponsors page with hero, principal, and gold tier sections
