+++ src/sections/oursponsors/OurSponsorsHeroSection.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export const OurSponsorsHeroSection: React.FC = () => {
+  return (
+    <section className="os-hero u-align--center">
+      <div className="row">
+        <div className="col-12">
+          <span className="os-hero__eyebrow">Sponsors &amp; Partners</span>
+          <h1 className="os-hero__title">Our Sponsors</h1>
+          <p className="os-hero__lead">
+            Acknowledging the organizations that make UbuCon India 2026 possible.
+          </p>
+        </div>
+      </div>
+    </section>
+  );
+};
