+++ src/styles/OurSponsors.scss
@@ -1,36 +1,14 @@
 // ─────────────────────────────────────────────────────────────────────────────
 // oursponsors.scss
 // All styles for the Our Sponsors page.
-// Hover colour-reveal is handled entirely in CSS (no inline onMouseEnter).
 // ─────────────────────────────────────────────────────────────────────────────
 
-// ── Local tokens ─────────────────────────────────────────────────────────────
 $os-brand:            var(--vf-color-brand);
 $os-accent:           var(--vf-color-accent);
-$os-bg-subtle:        #fcfcfc;
-$os-bg-tint-orange:   rgba(233, 84, 32, 0.04);
-$os-border-default:   #eaeaea;
-$os-border-subtle:    #e0e0e0;
-$os-border-orange:    rgba(233, 84, 32, 0.1);
-$os-text-muted:       #555;
-$os-text-faint:       #888;
-
-// Tier accent colours
-$os-gold:             #c9960c;
-$os-gold-bg:          rgba(201, 150, 12, 0.07);
-$os-principal:        var(--vf-color-accent);
-$os-principal-bg:     rgba(119, 41, 83, 0.06);
-
-$os-card-radius:      8px;
-
-
-// ─────────────────────────────────────────────────────────────────────────────
-// 1. Hero
-// ─────────────────────────────────────────────────────────────────────────────
 
 .os-hero {
-  background-color: $os-bg-subtle;
-  border-bottom: 1px solid $os-border-default;
+  background-color: #fcfcfc;
+  border-bottom: 1px solid #eaeaea;
   padding: 4rem 0 3.5rem;
 
   &__eyebrow {
@@ -57,17 +35,12 @@ $os-card-radius:      8px;
   }
 
   &__lead {
-    color: $os-text-muted;
+    color: #555;
     font-size: 1.125rem;
     max-width: 52ch;
-    margin: 0;
+    margin: 0 auto; // Center it
   }
 }
-
-
-// ─────────────────────────────────────────────────────────────────────────────
-// 2. Shared tier header (label + heading + sub used in every section)
-// ─────────────────────────────────────────────────────────────────────────────
 
 .os-tier-header {
   text-align: center;
@@ -84,14 +57,14 @@ $os-card-radius:      8px;
     text-transform: uppercase;
 
     &--principal {
-      background-color: $os-principal-bg;
+      background-color: rgba(119, 41, 83, 0.06);
       color: $os-accent;
       border: 1px solid rgba(119, 41, 83, 0.2);
     }
 
     &--gold {
-      background-color: $os-gold-bg;
-      color: $os-gold;
+      background-color: rgba(201, 150, 12, 0.07);
+      color: #c9960c;
       border: 1px solid rgba(201, 150, 12, 0.25);
     }
   }
@@ -104,7 +77,7 @@ $os-card-radius:      8px;
   }
 
   &__sub {
-    color: $os-text-muted;
+    color: #555;
     font-size: 0.9375rem;
     max-width: 56ch;
     margin: 0 auto;
@@ -112,84 +85,68 @@ $os-card-radius:      8px;
   }
 }
 
-
-// ─────────────────────────────────────────────────────────────────────────────
-// 3. Principal Sponsor section
-// ─────────────────────────────────────────────────────────────────────────────
-
 .os-principal-section {
   background-color: #ffffff;
+  padding: 4rem 0;
 }
-
-
-// ─────────────────────────────────────────────────────────────────────────────
-// 4. Gold Sponsors section
-// ─────────────────────────────────────────────────────────────────────────────
 
 .os-gold-section {
-  background-color: $os-bg-tint-orange;
-  border-top: 1px solid $os-border-orange;
-  border-bottom: 1px solid $os-border-orange;
-  padding: 3.5rem 0;
+  background-color: rgba(233, 84, 32, 0.04);
+  border-top: 1px solid rgba(233, 84, 32, 0.1);
+  border-bottom: 1px solid rgba(233, 84, 32, 0.1);
+  padding: 4rem 0;
 }
 
-
-// ─────────────────────────────────────────────────────────────────────────────
-// 5. Logo grid layouts
-// ─────────────────────────────────────────────────────────────────────────────
-
-.os-logos {
+/* Flexbox Tier Layout matching Home Screen */
+.os-tier-container {
   display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 1.5rem;
-
-  // Principal: single large centred card
-  &--principal {
-    gap: 0;
-  }
-
-  // Gold: row of medium cards
-  &--gold {
-    gap: 1.5rem;
-  }
+  flex-direction: column;
+  align-items: center;
 }
 
+.os-logos-grid {
+  display: flex;
+  gap: 1.5rem;
+  justify-content: center;
+  flex-wrap: wrap;
+  width: 100%;
+  max-width: 900px;
+}
 
-// ─────────────────────────────────────────────────────────────────────────────
-// 6. Logo card
-// ─────────────────────────────────────────────────────────────────────────────
+.os-logos-principal {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+}
 
 .os-logo-card {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  gap: 0.75rem;
   background-color: #ffffff;
-  border: 1px solid $os-border-default;
-  border-radius: $os-card-radius;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.03);
+  border: 1px solid #eaeaea;
+  border-radius: 6px;
+  box-shadow: 0 4px 12px rgba(0,0,0,0.02);
   cursor: default;
-  transition: box-shadow 0.18s ease, border-color 0.18s ease;
+  transition: box-shadow 0.2s ease, border-color 0.2s ease, transform 0.2s ease;
+  padding: 1.5rem;
+  width: 100%;
 
-  // Size variants
   &--lg {
-    width: 300px;
+    max-width: 300px;
     height: 130px;
-    padding: 1.75rem 2rem;
   }
 
   &--md {
-    width: 220px;
+    max-width: 240px;
     height: 100px;
-    padding: 1.25rem 1.5rem;
   }
 
-  // Hover: lift + reveal logo colour
   &:hover {
-    box-shadow: 0 6px 20px rgba(0, 0, 0, 0.08);
-    border-color: $os-border-subtle;
+    box-shadow: 0 8px 24px rgba(0, 0, 0, 0.08);
+    border-color: #d0d0d0;
+    transform: translateY(-4px);
 
     .os-logo-card__img {
       filter: none;
@@ -201,68 +158,47 @@ $os-card-radius:      8px;
     }
   }
 
-  // ── Logo image ────────────────────────────────────────────────────────────
-
   &__img {
-    max-width: 80%;
-    max-height: 44px;
+    max-height: 48px;
+    max-width: 85%;
     object-fit: contain;
-    // Default: desaturated, slightly faded
-    filter: grayscale(100%) opacity(0.7);
-    opacity: 0.85;
-    transition: filter 0.2s ease, opacity 0.2s ease;
+    filter: grayscale(100%) opacity(0.8);
+    transition: all 0.2s ease;
 
     .os-logo-card--lg & {
-      max-height: 52px;
+      max-height: 60px;
     }
   }
-
-  // ── Sponsor name (shows on hover as a subtle label) ───────────────────────
 
   &__name {
     font-size: 0.72rem;
     font-weight: 600;
     letter-spacing: 0.04em;
     text-transform: uppercase;
-    color: $os-text-faint;
+    color: #888;
     opacity: 0;
-    transition: opacity 0.18s ease;
+    transition: opacity 0.2s ease;
+    margin-top: 0.75rem;
   }
 }
 
-
-// ─────────────────────────────────────────────────────────────────────────────
-// 7. Responsive overrides
-// ─────────────────────────────────────────────────────────────────────────────
-
+// Ensure responsive row behavior
 @media (max-width: 772px) {
   .os-hero {
     &__title { font-size: 2rem; }
+    text-align: center;
   }
-
+  
   .os-logo-card {
-    // On touch devices, always show logos in colour (no hover state)
     &__img {
       filter: none;
       opacity: 1;
     }
-
     &__name {
       opacity: 1;
     }
-
-    &--lg {
-      width: 100%;
-      max-width: 300px;
-      height: auto;
-      padding: 1.5rem;
-    }
-
     &--md {
-      width: calc(50% - 0.75rem);
-      height: auto;
-      min-height: 80px;
-      padding: 1rem;
+      max-width: calc(50% - 0.75rem);
     }
   }
 }
